requirements.txt
Pillow==8.2.0
tensorflow==2.5.0
tkinter==8.6
